more database implementation

+++ backend/src/main/java/com/waremaster/backend/repository/ProductMovementRequestRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.waremaster.backend.entity.ProductMovementRequest;
 
 @Repository
-public interface ProductMovementRequestRepository extends JpaRepository<ProductMovementRequest, Integer>{
+public interface ProductMovementRequestRepository extends JpaRepository<ProductMovementRequest, Long>{
 
 }
